Move expect_column_values_to_equal_three.py to a fixtures directory
