+++ tests/expectations/test_generate_diagnostic_checklist.py
@@ -1,6 +1,6 @@
 import json
 
-from .expect_column_values_to_equal_three import (
+from .fixtures.expect_column_values_to_equal_three import (
     ExpectColumnValuesToEqualThree,
     ExpectColumnValuesToEqualThree__SecondIteration,
     ExpectColumnValuesToEqualThree__ThirdIteration,
