+++ tests/expectations/test_run_diagnostics.py
@@ -9,7 +9,7 @@ from great_expectations.expectations.expectation import (
 )
 from great_expectations.expectations.registry import _registered_expectations
 
-from .expect_column_values_to_equal_three import (
+from .fixtures.expect_column_values_to_equal_three import (
     ExpectColumnValuesToEqualThree,
     ExpectColumnValuesToEqualThree__SecondIteration,
 )
